+++ app/src/main/java/com/example/learning_2/dao/VacationDao.kt
@@ -16,10 +16,10 @@ interface VacationDao {
 
     // Retrieve vacations by their unique IDs
     @Query("SELECT * FROM Vacation WHERE id IN (:vacationIds)")
-    suspend fun loadAllByIds(vacationIds: IntArray): List<Vacation>
+    suspend fun loadAllByIds(vacationIds: LongArray): List<Vacation>
 
     @Query("SELECT * FROM Vacation WHERE id = :vacationId LIMIT 1")
-    suspend fun getById(vacationId: Int): Vacation?
+    suspend fun getById(vacationId: Long): Vacation?
 
     // Find a vacation by its name
     @Query("SELECT * FROM Vacation WHERE title LIKE :name LIMIT 1")
@@ -27,20 +27,20 @@ interface VacationDao {
 
     // Insert multiple vacations
     @Insert
-    suspend fun insertAll(vararg vacations: Vacation) : List<Long>
+    suspend fun insertAll(vararg vacations: Vacation): List<Long>
 
     // Delete a specific vacation
     @Delete
     suspend fun delete(vacation: Vacation)
 
     @Query("SELECT COUNT(*) FROM Excursion WHERE vacation_id = :vacationId")
-    suspend fun countExcursionsForVacation(vacationId: Int): Int
+    suspend fun countExcursionsForVacation(vacationId: Long): Long
 
     @Update
     suspend fun update(selectedVacation: Vacation)
 
     // Check if a vacation has associated excursions
     @Query("SELECT COUNT(*) FROM excursion WHERE vacation_id = :vacationId")
-    suspend fun getExcursionCountForVacation(vacationId: Int): Int
+    suspend fun getExcursionCountForVacation(vacationId: Long): Long
 
 }
